created ShortestPathFasterAlgorithm.cpp
Approach: The shortest path faster algorithm is based on Bellman-Ford algorithm where every vertex is used to relax its adjacent vertices but in SPF algorithm, a queue of vertices is maintained and a vertex is added to the queue only if that vertex is relaxed. This process repeats until no more vertex can be relaxed. 
The following steps can be performed to compute the result: 
 

Create an array d[] to store the shortest distance of all vertex from the source vertex. Initialize this array by infinity except for d[S] = 0 where S is the source vertex.
Create a queue Q and push starting source vertex in it. 
while Queue is not empty, do the following for each edge(u, v) in the graph 
If d[v] > d[u] + weight of edge(u, v)
d[v] = d[u] + weight of edge(u, v)
If vertex v is not present in Queue, then push the vertex v into the Queue.
Note: The term relaxation means updating the cost of all vertices connected to a vertex v if those costs would be improved by including the path via vertex v. This can be understood better from an analogy between the estimate of the shortest path and the length of a helical tension spring, which is not designed for compression. Initially, the cost of the shortest path is an overestimate, likened to a stretched-out spring. As shorter paths are found, the estimated cost is lowered, and the spring is relaxed. Eventually, the shortest path, if one exists, is found and the spring has been relaxed to its resting length.
Below is the implementation of the above approach: